Application-client/requirements.txt
py_solc_x==1.1.1
web3==5.27.0
